// src/interfaces/auth.interface.ts
// ISignInData interface represents the data required for user sign-in
export interface ISignInData {
    username: string;
    password: string;
}


// ISignUpData interface represents the data required for user sign-up
export interface ISignUpData {
    username: string;
    first_name: string;
    last_name: string;
    email: string;
    password: string;
}